+++ packages/form-types/src/core/value-type.ts
@@ -5,7 +5,7 @@ export type OptionGroup = { label: string, options: SelectOption[] }
 export type SelectOption = string | Option | OptionGroup;
 
 export type ValueType<T extends keyof FieldConfigMap> =
-  T extends 'checkbox'
+  T extends 'checkbox' | 'switch'
     ? boolean
     : T extends 'number' | 'range'
       ? number