+++ packages/form-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0
+
+### Minor Changes
+
+- Feat: Add complete switch
+
 ## 0.2.0
 
 ### Minor Changes