+++ packages/form-validation/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @sio-group/form-types@0.3.0
+
 ## 0.1.4
 
 ### Patch Changes